requirements.txt
streamlit==1.35.0
mediapipe==0.10.20
protobuf==4.25.3
jieba==0.42.1
google-cloud-texttospeech==2.19.0
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
rapidfuzz==3.6.2